pyproject.toml
[project]
name = "trading_app"
version = "0.1.0"
description = ""
authors = [
    {name = "sl.ars",email = "sl.ars@icloud.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "stripe (>=11.5.0,<12.0.0)",
    "drf-yasg (>=1.21.8,<2.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
